Commit 4

+++ Assets/Scripts/test.cs
@@ -9,15 +9,22 @@ public class test : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        dynamicInventory.AddItem(item1);
-        dynamicInventory.AddItem(item1);
-        dynamicInventory.DeleteItem(item1);
-        dynamicInventory.FindItem(item1);
+        //dynamicInventory.AddItem(item1);
+        //dynamicInventory.AddItem(item1);
+        //dynamicInventory.DeleteItem(item1);
+        //dynamicInventory.FindItem(item1);
     }
 
     // Update is called once per frame
     void Update()
     {
-        
+        if (Input.GetKeyDown(KeyCode.Alpha1))
+        {
+            dynamicInventory.AddItem(item1);
+        }
+        if (Input.GetKeyDown(KeyCode.Alpha2))
+        {
+            dynamicInventory.DeleteItem(item1);
+        }
     }
 }
